Added script for handling database import from production environment to a secondary instance.
In upgrade script added variables for help desk url and apache user/group to make the script more modular and corrected an issue where we were failing to set file and directory permissions prior to calling for a cache rebuild when user elected to install TinyMCE patch.

Updated Readme documentation to reflect current.